* issue #28: auxiliary variable __retval introduced for functions that have nested functions.
Note: a RETURN expression may use nested functions and therefore the static link must be reset after the RETURN, and therfore an auxiliary variable is needed.
In addition it may use elements of a copied dynamic array (DUP), which are to be freed in DEL. Thiose are currently ignored because in Linux/gcc DEL is empty, i.e. not needed.